+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [4.0.0] 2025-09-18
+### Fixed
+- Prevent the dialog from closing before validation is complete.
+### Changed
+- Updated variable title to display Form or Item Variables.
+- Updated empty variable list message.
+- Added support for enableWhen expression.
+
 ## [3.3.2] 2025-08-26
 ### Fixed
 - Addressed styling issues affecting the help dialog.